Using URI string comparison to check for playability
